Added new TaskContext object for agents
Agent commands now can be cancelled via a TaskContext object. This object also lets commands increment the step counter and return new thoughts to the caller. A new `max_time` parameter lets you specify how long an Agent is allowed to perform a task. The default is 60000 ms or 1 minute.

This checkin also contains a couple of minor bug fixes and I published a new 0.7.0 version of all packages.